+++ rpc/backend.go
@@ -47,13 +47,15 @@ func NewEthermintBackend(cliCtx context.CLIContext) *EthermintBackend {
 
 // BlockNumber returns the current block number.
 func (e *EthermintBackend) BlockNumber() (hexutil.Uint64, error) {
-	res, _, err := e.cliCtx.QueryWithData(fmt.Sprintf("custom/%s/blockNumber", types.ModuleName), nil)
+	res, height, err := e.cliCtx.QueryWithData(fmt.Sprintf("custom/%s/blockNumber", types.ModuleName), nil)
 	if err != nil {
 		return hexutil.Uint64(0), err
 	}
 
 	var out types.QueryResBlockNumber
 	e.cliCtx.Codec.MustUnmarshalJSON(res, &out)
+
+	e.cliCtx.WithHeight(height)
 	return hexutil.Uint64(out.Number), nil
 }